Templating TimeSteppers
