+++ private/julio/mesh_free/G.Fasshauer/1.2_distance_matrix/demo_distance_matrix.cpp
@@ -10,7 +10,7 @@ using namespace scicellxx;
 template<class MAT_TYPE, class VEC_TYPE>
 class Problem
 {
-oublic:
+public:
  Problem();
  ~Problem();
