+++ src/data_structures/data_structures.h
@@ -3,7 +3,6 @@
 
 #include "../general/general.h"
 
-#include "ac_odes.h"
 #include "cc_data.h"
 #include "cc_node.h"
 #include "cc_boundary_node.h"
