+++ src/equations/equations.h
@@ -1,6 +1,7 @@
 #ifndef SCICELLXX_EQUATIONS_H
 #define SCICELLXX_EQUATIONS_H
 
+#include "ac_odes.h"
 #include "ac_jacobian_and_residual.h"
 
 #endif // #ifndef SCICELLXX_EQUATIONS_H
