+++ src/time_steppers/cc_adams_moulton_2_method.tpl.cpp
@@ -1,4 +1,4 @@
-#include "cc_bdf_2_method.h"
+#include "cc_adams_moulton_2_method.tpl.h"
 
 namespace scicellxx
 {
@@ -6,15 +6,14 @@ namespace scicellxx
  // ===================================================================
  /// Constructor
  // ===================================================================
- CCBDF2Method::CCBDF2Method()
-  : ACTimeStepperForODEs(),
-    Compute_u_at_time_t_plus_h(true)
- {
+ template<class EQUATIONS_TYPE>
+ CCAdamsMoulton2Method<EQUATIONS_TYPE>::CCAdamsMoulton2Method()
+  : ACTimeStepper<EQUATIONS_TYPE>()
+ {  
   // Sets the number of history values
-  N_history_values = 3;
+  this->N_history_values = 2;
   
-  //Newtons_method.set_newton_absolute_solver_tolerance(1.0e-3);
-  //Newtons_method.set_maximum_allowed_residual(1.0e-1);
+  //Newtons_method.set_newton_solver_tolerance(1.0e-3);
   
   // Disable output for Newton's method and relative tolerance
   Newtons_method.disable_output_messages();
@@ -24,33 +23,36 @@ namespace scicellxx
  // ===================================================================
  /// Empty destructor
  // ===================================================================
- CCBDF2Method::~CCBDF2Method()
+ template<class EQUATIONS_TYPE>
+ CCAdamsMoulton2Method<EQUATIONS_TYPE>::~CCAdamsMoulton2Method()
  {
   
  }
  
  // ===================================================================
- /// Applies BDF2 method to the given odes from the current time "t" to
- /// the time "t+h". The values of u at time t+h will be stored at
- /// index k (default k = 0).
- // ===================================================================
- void CCBDF2Method::time_step(ACODEs &odes, const Real h,
-                              const Real t,
-                              CCData &u,
-                              const unsigned k)
+ /// Applies Adams-Moulton 2 method to the given odes from the current
+ /// time "t" to the time "t+h". The values of u at time t+h will be /
+ //stored at index k (default k = 0).
+ //===================================================================
+ template<class EQUATIONS_TYPE>
+ void CCAdamsMoulton2Method<EQUATIONS_TYPE>::time_step(EQUATIONS_TYPE &odes,
+                                                       const Real h,
+                                                       const Real t,
+                                                       CCData &u,
+                                                       const unsigned k)
  {
 #ifdef SCICELLXX_PANIC_MODE
   // Check if the ode has the correct number of history values to
-  // apply Backward-Euler's method
+  // apply Euler's method
   const unsigned n_history_values = u.n_history_values();
-  if (n_history_values < N_history_values)
+  if (n_history_values < this->N_history_values)
    {
     // Error message
     std::ostringstream error_message;
     error_message << "The number of history values is less than\n"
-                  << "the required by Backward Euler's method" << std::endl
+                  << "the required by Adams-Moulton 2 method" << std::endl
                   << "Required number of history values: "
-                  << N_history_values << std::endl
+                  << this->N_history_values << std::endl
                   << "Number of history values: "
                   << n_history_values << std::endl;
     throw SciCellxxLibError(error_message.str(),
@@ -58,29 +60,13 @@ namespace scicellxx
                            SCICELLXX_EXCEPTION_LOCATION);
    }
 #endif // #ifdef SCICELLXX_PANIC_MODE
-  
-  // -----------------------------------------------------------------
-  // Compute the value of u_{t+h} if this is the first time
-  if (Compute_u_at_time_t_plus_h)
-   {
-    // Compute the values for u_{t+h} using the same time stepper to
-    // compute the initial guess for Newton's method
-    Time_stepper_initial_guess.time_step(odes, h, t, u, k);
     
-    // This should be performed only once
-    Compute_u_at_time_t_plus_h = false;
-    
-    // Return. The next time we will have the required values for BDF2
-    // u_{t} and u_{t+h}
-    return;
-   }
-  
   // -----------------------------------------------------------------
   // Compute initial guess
   // -----------------------------------------------------------------  
   // Compute the initial guess for Newton's method using the values of
   // u at time 't', the values of u at time 't+h' are automatically
-  // shifted at index k
+  // shifted at index k 
   Time_stepper_initial_guess.time_step(odes, h, t, u, k);
   
   // ---------------------------------------------------
@@ -88,9 +74,10 @@ namespace scicellxx
   // ---------------------------------------------------
   // Get the number of odes
   const unsigned n_odes = odes.n_odes();
-  
+
   // Create a vector with the initial guess from the first row (0)
   // since the values have been shift
+
 #ifdef SCICELLXX_USES_ARMADILLO
   CCVectorArmadillo u_initial_guess(u.history_values_row_pt(0), n_odes);
 #else
@@ -109,8 +96,8 @@ namespace scicellxx
   // Solve using Newton's method, the solution is automatically copied
   // back at the u data structure
   Newtons_method.solve(&u_initial_guess);
-  
+    
  }
- 
+
 }
