+++ src/time_steppers/cc_backward_euler_method.tpl.cpp
@@ -1,4 +1,4 @@
-#include "cc_adams_moulton_2_method.h"
+#include "cc_backward_euler_method.tpl.h"
 
 namespace scicellxx
 {
@@ -6,13 +6,15 @@ namespace scicellxx
  // ===================================================================
  /// Constructor
  // ===================================================================
- CCAdamsMoulton2Method::CCAdamsMoulton2Method()
-  : ACTimeStepperForODEs()
+ template<class EQUATIONS_TYPE>
+ CCBackwardEulerMethod<EQUATIONS_TYPE>::CCBackwardEulerMethod()
+  : ACTimeStepper<EQUATIONS_TYPE>()
  {  
   // Sets the number of history values
   this->N_history_values = 2;
   
-  //Newtons_method.set_newton_solver_tolerance(1.0e-3);
+  //Newtons_method.set_newton_absolute_solver_tolerance(1.0e-3);
+  //Newtons_method.set_maximum_newton_iterations(100);
   
   // Disable output for Newton's method and relative tolerance
   Newtons_method.disable_output_messages();
@@ -22,34 +24,36 @@ namespace scicellxx
  // ===================================================================
  /// Empty destructor
  // ===================================================================
- CCAdamsMoulton2Method::~CCAdamsMoulton2Method()
+ template<class EQUATIONS_TYPE>
+ CCBackwardEulerMethod<EQUATIONS_TYPE>::~CCBackwardEulerMethod()
  {
   
  }
  
  // ===================================================================
- /// Applies Adams-Moulton 2 method to the given odes from the current
- /// time "t" to the time "t+h". The values of u at time t+h will be /
- //stored at index k (default k = 0).
- //===================================================================
- void CCAdamsMoulton2Method::time_step(ACODEs &odes,
-                                       const Real h,
-                                       const Real t,
-                                       CCData &u,
-                                       const unsigned k)
+ /// Applies Backward Euler method to the given odes from the current
+ /// time "t" to the time "t+h". The values of u at time t+h will be
+ /// stored at index k (default k = 0).
+ // ===================================================================
+ template<class EQUATIONS_TYPE>
+ void CCBackwardEulerMethod<EQUATIONS_TYPE>::time_step(EQUATIONS_TYPE &odes,
+                                                       const Real h,
+                                                       const Real t,
+                                                       CCData &u,
+                                                       const unsigned k)
  {
 #ifdef SCICELLXX_PANIC_MODE
   // Check if the ode has the correct number of history values to
-  // apply Euler's method
+  // apply Backward-Euler's method
   const unsigned n_history_values = u.n_history_values();
-  if (n_history_values < N_history_values)
+  if (n_history_values < this->N_history_values)
    {
     // Error message
     std::ostringstream error_message;
     error_message << "The number of history values is less than\n"
-                  << "the required by Adams-Moulton 2 method" << std::endl
+                  << "the required by Backward Euler's method" << std::endl
                   << "Required number of history values: "
-                  << N_history_values << std::endl
+                  << this->N_history_values << std::endl
                   << "Number of history values: "
                   << n_history_values << std::endl;
     throw SciCellxxLibError(error_message.str(),
@@ -57,7 +61,7 @@ namespace scicellxx
                            SCICELLXX_EXCEPTION_LOCATION);
    }
 #endif // #ifdef SCICELLXX_PANIC_MODE
-    
+  
   // -----------------------------------------------------------------
   // Compute initial guess
   // -----------------------------------------------------------------  
@@ -71,10 +75,9 @@ namespace scicellxx
   // ---------------------------------------------------
   // Get the number of odes
   const unsigned n_odes = odes.n_odes();
-
+  
   // Create a vector with the initial guess from the first row (0)
   // since the values have been shift
-
 #ifdef SCICELLXX_USES_ARMADILLO
   CCVectorArmadillo u_initial_guess(u.history_values_row_pt(0), n_odes);
 #else
@@ -93,8 +96,8 @@ namespace scicellxx
   // Solve using Newton's method, the solution is automatically copied
   // back at the u data structure
   Newtons_method.solve(&u_initial_guess);
-    
+  
  }
-
+ 
 }
