+++ src/time_steppers/cc_bdf_2_method.tpl.cpp
@@ -1,4 +1,4 @@
-#include "cc_backward_euler_method.h"
+#include "cc_bdf_2_method.tpl.h"
 
 namespace scicellxx
 {
@@ -6,14 +6,16 @@ namespace scicellxx
  // ===================================================================
  /// Constructor
  // ===================================================================
- CCBackwardEulerMethod::CCBackwardEulerMethod()
-  : ACTimeStepperForODEs()
- {  
+ template<class EQUATIONS_TYPE>
+ CCBDF2Method<EQUATIONS_TYPE>::CCBDF2Method()
+  : ACTimeStepper<EQUATIONS_TYPE>(),
+    Compute_u_at_time_t_plus_h(true)
+ {
   // Sets the number of history values
-  this->N_history_values = 2;
+  this->N_history_values = 3;
   
   //Newtons_method.set_newton_absolute_solver_tolerance(1.0e-3);
-  //Newtons_method.set_maximum_newton_iterations(100);
+  //Newtons_method.set_maximum_allowed_residual(1.0e-1);
   
   // Disable output for Newton's method and relative tolerance
   Newtons_method.disable_output_messages();
@@ -23,34 +25,35 @@ namespace scicellxx
  // ===================================================================
  /// Empty destructor
  // ===================================================================
- CCBackwardEulerMethod::~CCBackwardEulerMethod()
+ template<class EQUATIONS_TYPE>
+ CCBDF2Method<EQUATIONS_TYPE>::~CCBDF2Method()
  {
   
  }
  
  // ===================================================================
- /// Applies Backward Euler method to the given odes from the current
- /// time "t" to the time "t+h". The values of u at time t+h will be
- /// stored at index k (default k = 0).
+ /// Applies BDF2 method to the given odes from the current time "t" to
+ /// the time "t+h". The values of u at time t+h will be stored at
+ /// index k (default k = 0).
  // ===================================================================
- void CCBackwardEulerMethod::time_step(ACODEs &odes,
-                                       const Real h,
-                                       const Real t,
-                                       CCData &u,
-                                       const unsigned k)
+ template<class EQUATIONS_TYPE>
+ void CCBDF2Method<EQUATIONS_TYPE>::time_step(EQUATIONS_TYPE &odes, const Real h,
+                                              const Real t,
+                                              CCData &u,
+                                              const unsigned k)
  {
 #ifdef SCICELLXX_PANIC_MODE
   // Check if the ode has the correct number of history values to
   // apply Backward-Euler's method
   const unsigned n_history_values = u.n_history_values();
-  if (n_history_values < N_history_values)
+  if (n_history_values < this->N_history_values)
    {
     // Error message
     std::ostringstream error_message;
     error_message << "The number of history values is less than\n"
                   << "the required by Backward Euler's method" << std::endl
                   << "Required number of history values: "
-                  << N_history_values << std::endl
+                  << this->N_history_values << std::endl
                   << "Number of history values: "
                   << n_history_values << std::endl;
     throw SciCellxxLibError(error_message.str(),
@@ -60,11 +63,27 @@ namespace scicellxx
 #endif // #ifdef SCICELLXX_PANIC_MODE
   
   // -----------------------------------------------------------------
+  // Compute the value of u_{t+h} if this is the first time
+  if (Compute_u_at_time_t_plus_h)
+   {
+    // Compute the values for u_{t+h} using the same time stepper to
+    // compute the initial guess for Newton's method
+    Time_stepper_initial_guess.time_step(odes, h, t, u, k);
+    
+    // This should be performed only once
+    Compute_u_at_time_t_plus_h = false;
+    
+    // Return. The next time we will have the required values for BDF2
+    // u_{t} and u_{t+h}
+    return;
+   }
+  
+  // -----------------------------------------------------------------
   // Compute initial guess
   // -----------------------------------------------------------------  
   // Compute the initial guess for Newton's method using the values of
   // u at time 't', the values of u at time 't+h' are automatically
-  // shifted at index k 
+  // shifted at index k
   Time_stepper_initial_guess.time_step(odes, h, t, u, k);
   
   // ---------------------------------------------------
