+++ src/time_steppers/cc_jacobian_and_residual_for_backward_euler.tpl.cpp
@@ -1,34 +1,37 @@
-#include "cc_jacobian_and_residual_for_bdf_2.h"
+#include "cc_jacobian_and_residual_for_backward_euler.tpl.h"
 
 namespace scicellxx
 {
  // ===================================================================
  /// Constructor
  // ===================================================================
- CCJacobianAndResidualForBDF2::CCJacobianAndResidualForBDF2()
-  : ACJacobianAndResidualForImplicitTimeStepper()
+ template<class EQUATIONS_TYPE>
+ CCJacobianAndResidualForBackwardEuler<EQUATIONS_TYPE>::CCJacobianAndResidualForBackwardEuler()
+  : ACJacobianAndResidualForImplicitTimeStepper<EQUATIONS_TYPE>()
  {
   // Assign the default strategy (the Finite Differences strategy) for
   // computation of the Jacobian of the ODEs
-  this->set_strategy_for_odes_jacobian(&Jacobian_by_FD_strategy);  
+  this->set_strategy_for_odes_jacobian(&Jacobian_by_FD_strategy);
  }
  
  // ===================================================================
  /// Empty destructor
  // ===================================================================
- CCJacobianAndResidualForBDF2::~CCJacobianAndResidualForBDF2()
+ template<class EQUATIONS_TYPE>
+ CCJacobianAndResidualForBackwardEuler<EQUATIONS_TYPE>::~CCJacobianAndResidualForBackwardEuler()
  {
   
  }
  
  // ===================================================================
- /// In charge of computing the Jacobian (virtual function
- /// implementation)
- //===================================================================
- void CCJacobianAndResidualForBDF2::compute_jacobian()
+ /// In charge of computing the Jacobian
+ /// (virtual function implementation)
+ // ===================================================================
+ template<class EQUATIONS_TYPE>
+ void CCJacobianAndResidualForBackwardEuler<EQUATIONS_TYPE>::compute_jacobian()
  {
   // Get the odes
-  ACODEs *odes_pt = this->odes_pt();
+  EQUATIONS_TYPE *odes_pt = this->odes_pt();
   // Get the time step
   const Real h = this->time_step();
   // Get the current time
@@ -57,9 +60,8 @@ namespace scicellxx
   // -------------------------------------------
   // Compute the Jacobian of F(Y) at time t+h
   // -------------------------------------------
-
   // Get a pointer to the strategy to compute the Jacobian of the ODEs
-  ACJacobianAndResidualForImplicitTimeStepper *jacobian_strategy_odes_pt =
+  ACJacobianAndResidualForImplicitTimeStepper<EQUATIONS_TYPE> *jacobian_strategy_odes_pt =
    this->jacobian_FY_strategy_pt();
   
   // Set the data for the computation of the jacobian and the residual
@@ -78,21 +80,18 @@ namespace scicellxx
   // Allocate memory for the Jacobian (delete previous data)
   this->Jacobian_pt->allocate_memory(n_dof, n_dof);
   
-  // Two thirds of the time step
-  const Real h_two_thirds = (2.0/3.0) * h;
-  
-  // Compute the approximated Jacobian (I - \frac{2}{3}h * Jacobian_FY(i, j))
+  // Compute the approximated Jacobian (I - h * Jacobian_FY(i, j))
   for (unsigned i = 0; i < n_dof; i++)
    {
     for (unsigned j = 0; j < n_dof; j++)
      {
       if (i == j)
        {
-        (*this->Jacobian_pt)(i, j) = 1.0 - (h_two_thirds * Jacobian_FY_pt->value(i, j));
+        (*this->Jacobian_pt)(i, j) = 1.0 - (h * Jacobian_FY_pt->value(i, j));
        }
       else
        {
-        (*this->Jacobian_pt)(i, j) = 0.0 - (h_two_thirds * Jacobian_FY_pt->value(i, j));
+        (*this->Jacobian_pt)(i, j) = 0.0 - (h * Jacobian_FY_pt->value(i, j));
        }
      }
    }
@@ -100,12 +99,13 @@ namespace scicellxx
  }
  
  // ===================================================================
- // In charge of computing the residual
+ /// In charge of computing the residual
  // ===================================================================
- void CCJacobianAndResidualForBDF2::compute_residual()
+ template<class EQUATIONS_TYPE>
+ void CCJacobianAndResidualForBackwardEuler<EQUATIONS_TYPE>::compute_residual()
  {
   // Get the odes
-  ACODEs *odes_pt = this->odes_pt();
+  EQUATIONS_TYPE *odes_pt = this->odes_pt();
   // Get the time step
   const Real h = this->time_step();
   // Get the current time
@@ -134,27 +134,22 @@ namespace scicellxx
   // Get the number of ODEs
   const unsigned n_dof = odes_pt->n_odes();
   
-  // Temporary vector to store the evaluation of the odes at time
-  // 't+h' (current Newton's iteration)
-  CCData dudt(n_dof);
+  // Temporary vector to store the evaluation of the odes
+  CCData dudt_new(n_dof);
   
-  // Time step weights
-  const Real h_two_thirds = (2.0/3.0) * h;
-  
-  // Evaluate the odes at time "t+h". Current Newton's iteration
-  odes_pt->evaluate_derivatives(t+h, (*u_pt), dudt, k);
+  // Evaluate the ODE at time 't+h', stored at index k
+  odes_pt->evaluate_derivatives(t+h, (*u_pt), dudt_new, k);
   
   // Allocate memory for the Residual (delete previous data)
   this->Residual_pt->allocate_memory(n_dof);
   
-  // F(Y) - (u_{t+h} - \frac{4}{3} u_{t} + \frac{1}{3} u_{t-h} - \frac{2}{3} h f(t+h,u_{t+h}))
+  // F(Y) = -(u_{t+h} - u_{t} - h f(t+h, u_{t+h}))
   for (unsigned i = 0; i < n_dof; i++)
    {
-    this->Residual_pt->value(i) =
-     -(u_pt->value(i,k) - ((4.0/3.0) * u_pt->value(i,k+1)) + ((1.0/3.0) * u_pt->value(i,k+2)) - (h_two_thirds * dudt(i)));
+    this->Residual_pt->value(i) = -(u_pt->value(i,k) - u_pt->value(i,k+1) - (h * dudt_new(i)));
    }
   
  }
- 
+  
 }
