+++ src/time_steppers/cc_jacobian_and_residual_for_bdf_2.tpl.cpp
@@ -1,24 +1,24 @@
-#include "cc_jacobian_and_residual_for_adams_moulton_2.h"
+#include "cc_jacobian_and_residual_for_bdf_2.tpl.h"
 
 namespace scicellxx
 {
  // ===================================================================
  /// Constructor
  // ===================================================================
- CCJacobianAndResidualForAdamsMoulton2::CCJacobianAndResidualForAdamsMoulton2()
-  : ACJacobianAndResidualForImplicitTimeStepper(),
-    DUDT_old(NULL),
-    Evaluate_odes_with_u_old_values(false)
+ template<class EQUATIONS_TYPE>
+ CCJacobianAndResidualForBDF2<EQUATIONS_TYPE>::CCJacobianAndResidualForBDF2()
+  : ACJacobianAndResidualForImplicitTimeStepper<EQUATIONS_TYPE>()
  {
   // Assign the default strategy (the Finite Differences strategy) for
   // computation of the Jacobian of the ODEs
-  this->set_strategy_for_odes_jacobian(&Jacobian_by_FD_strategy);
+  this->set_strategy_for_odes_jacobian(&Jacobian_by_FD_strategy);  
  }
  
  // ===================================================================
  /// Empty destructor
  // ===================================================================
- CCJacobianAndResidualForAdamsMoulton2::~CCJacobianAndResidualForAdamsMoulton2()
+ template<class EQUATIONS_TYPE>
+ CCJacobianAndResidualForBDF2<EQUATIONS_TYPE>::~CCJacobianAndResidualForBDF2()
  {
   
  }
@@ -27,10 +27,11 @@ namespace scicellxx
  /// In charge of computing the Jacobian (virtual function
  /// implementation)
  //===================================================================
- void CCJacobianAndResidualForAdamsMoulton2::compute_jacobian()
+ template<class EQUATIONS_TYPE>
+ void CCJacobianAndResidualForBDF2<EQUATIONS_TYPE>::compute_jacobian()
  {
   // Get the odes
-  ACODEs *odes_pt = this->odes_pt();
+  EQUATIONS_TYPE *odes_pt = this->odes_pt();
   // Get the time step
   const Real h = this->time_step();
   // Get the current time
@@ -61,14 +62,14 @@ namespace scicellxx
   // -------------------------------------------
   
   // Get a pointer to the strategy to compute the Jacobian of the ODEs
-  ACJacobianAndResidualForImplicitTimeStepper *jacobian_strategy_odes_pt =
+  ACJacobianAndResidualForImplicitTimeStepper<EQUATIONS_TYPE> *jacobian_strategy_odes_pt =
    this->jacobian_FY_strategy_pt();
   
   // Set the data for the computation of the jacobian and the residual
   jacobian_strategy_odes_pt->set_data_for_jacobian_and_residual(odes_pt, h, t, u_pt, k);
   
   // Compute Jacobian
-  jacobian_strategy_odes_pt->compute_jacobian(); 
+  jacobian_strategy_odes_pt->compute_jacobian();
   
   // Store the Jacobian for FY, used in the computation of the
   // backward Euler Jacobian $J = I - (h * Jacobian_{FY})$
@@ -80,21 +81,21 @@ namespace scicellxx
   // Allocate memory for the Jacobian (delete previous data)
   this->Jacobian_pt->allocate_memory(n_dof, n_dof);
   
-  // Half time step
-  const Real h_half = h * 0.5;
+  // Two thirds of the time step
+  const Real h_two_thirds = (2.0/3.0) * h;
   
-  // Compute the approximated Jacobian (I - h * Jacobian_FY(i, j))
+  // Compute the approximated Jacobian (I - \frac{2}{3}h * Jacobian_FY(i, j))
   for (unsigned i = 0; i < n_dof; i++)
    {
     for (unsigned j = 0; j < n_dof; j++)
      {
       if (i == j)
        {
-        (*this->Jacobian_pt)(i, j) = 1.0 - (h_half * Jacobian_FY_pt->value(i, j));
+        (*this->Jacobian_pt)(i, j) = 1.0 - (h_two_thirds * Jacobian_FY_pt->value(i, j));
        }
       else
        {
-        (*this->Jacobian_pt)(i, j) = 0.0 - (h_half * Jacobian_FY_pt->value(i, j));
+        (*this->Jacobian_pt)(i, j) = 0.0 - (h_two_thirds * Jacobian_FY_pt->value(i, j));
        }
      }
    }
@@ -102,12 +103,13 @@ namespace scicellxx
  }
  
  // ===================================================================
- /// In charge of computing the residual
+ // In charge of computing the residual
  // ===================================================================
- void CCJacobianAndResidualForAdamsMoulton2::compute_residual()
+ template<class EQUATIONS_TYPE>
+ void CCJacobianAndResidualForBDF2<EQUATIONS_TYPE>::compute_residual()
  {
   // Get the odes
-  ACODEs *odes_pt = this->odes_pt();
+  EQUATIONS_TYPE *odes_pt = this->odes_pt();
   // Get the time step
   const Real h = this->time_step();
   // Get the current time
@@ -131,70 +133,29 @@ namespace scicellxx
     throw SciCellxxLibError(error_message.str(),
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
-   } 
+   }
   
   // Get the number of ODEs
   const unsigned n_dof = odes_pt->n_odes();
   
-  // Is this the first time we called this method?
-  if (Evaluate_odes_with_u_old_values)
-   {
-    // Free memory if there is something in there
-    if (DUDT_old!= NULL)
-     {
-      delete DUDT_old;
-      DUDT_old = NULL;
-     }
-    
-    // Temporary vector to store the evaluation of the odes with u
-    // values at time 't' (constant values)
-    DUDT_old = new CCData(n_dof);
-    
-    // Evaluate the ODEs with the values of u at time "t". Constant
-    // during Newton's iteration
-    odes_pt->evaluate_derivatives(t, (*u_pt), (*DUDT_old), k+1);
-    
-    // Avoid evaluation of the same function during following Newton's
-    // iterations
-    Evaluate_odes_with_u_old_values = false;
-    
-   }
-  
-  // Check whether we have previously evaluated the derivative with
-  // the values of u at time 't'
-  if (DUDT_old==NULL)
-   {
-    // Error message
-    std::ostringstream error_message;
-    error_message << "You have not evaluated the odes with the values of u at time 't'\n"
-                  << "These values are required to compute the residual\n"
-                  << "Call the method\n"
-                  << "enable_evaluation_of_odes_with_old_u_values()\n"
-                  << "ONLY before the FIRST computation of the residual\n"
-                  << std::endl;
-    throw SciCellxxLibError(error_message.str(),
-                           SCICELLXX_CURRENT_FUNCTION,
-                           SCICELLXX_EXCEPTION_LOCATION);
-   }
-  
-  // Temporary vector to store the evaluation of the odes with the
-  // values of u at time 't+h' (current Newton's iteration)
+  // Temporary vector to store the evaluation of the odes at time
+  // 't+h' (current Newton's iteration)
   CCData dudt(n_dof);
   
-  // Evaluate the ODE at time "t+h"
+  // Time step weights
+  const Real h_two_thirds = (2.0/3.0) * h;
+  
+  // Evaluate the odes at time "t+h". Current Newton's iteration
   odes_pt->evaluate_derivatives(t+h, (*u_pt), dudt, k);
   
   // Allocate memory for the Residual (delete previous data)
   this->Residual_pt->allocate_memory(n_dof);
   
-  // Half time step
-  const Real h_half = h * 0.5;
-
-  // F(Y) = -(u_{t+h} - u_{t} - \frac{1}{2} (f(t+h, u_{t+h}) + f(t, u_{t})) )
+  // F(Y) - (u_{t+h} - \frac{4}{3} u_{t} + \frac{1}{3} u_{t-h} - \frac{2}{3} h f(t+h,u_{t+h}))
   for (unsigned i = 0; i < n_dof; i++)
    {
     this->Residual_pt->value(i) =
-     -(u_pt->value(i,k) - u_pt->value(i,k+1) - (h_half * (dudt(i) + DUDT_old->value(i))));
+     -(u_pt->value(i,k) - ((4.0/3.0) * u_pt->value(i,k+1)) + ((1.0/3.0) * u_pt->value(i,k+2)) - (h_two_thirds * dudt(i)));
    }
   
  }
